IGAPP-635: Move News and test
