+++ native/src/components/__tests__/News.spec.tsx
